refactor(api): Modularize API service layer
The single `api.ts` file was growing too large and becoming difficult to maintain, coupling unrelated concerns like authentication, chores, and expenses.

This refactoring breaks down the monolithic service into smaller, domain-specific modules for improved maintainability, readability, and scalability.

Key changes:
- Extracted all TypeScript interfaces into a central `src/lib/types/types.ts` file.
- Decomposed `api.ts` into separate files based on functionality (auth, chores, expenses, households, etc.) and placed them under `src/lib/api/`.
- Created a barrel file (`src/lib/api/index.ts`) to re-export all modules for cleaner, optional centralized imports.
- Updated all components (`RoomiesApp`, `ChoreDashboard`, `HouseholdSettings`, etc.) to use the new modular imports, removing the dependency on the old `api.ts`.
