+++ src/lib/api/index.ts
@@ -1,0 +1,10 @@
+// src/lib/api/index.ts
+export * from './auth';
+export * from './chores';
+export * from './expenses';
+export * from './households';
+export * from './messages';
+export * from './notifications';
+export * from './profile';
+export * from './settlements';
+export * from './tasks';
